move rc validation to its own file
