+++ src/cli/tc/mod.rs
@@ -4,3 +4,4 @@ mod helpers;
 mod init;
 pub mod run;
 mod start;
+mod validate_rc;
